+++ packages/dsw-data-seeder/dsw/data_seeder/cli.py
@@ -6,7 +6,7 @@ import typing
 
 import click
 
-from dsw.config.parser import MissingConfigurationError
+from dsw.config.parser import InvalidConfigurationError, MissingConfigurationError
 
 from . import consts
 from .config import SeederConfig, SeederConfigParser
@@ -15,13 +15,14 @@ from .seeder import DataSeeder, SeedRecipe, SentryReporter
 
 def load_config_str(config_str: str) -> SeederConfig:
     parser = SeederConfigParser()
-    if not parser.can_read(config_str):
-        click.echo('Error: Cannot parse config file', err=True)
-        sys.exit(1)
 
     try:
         parser.read_string(config_str)
         parser.validate()
+    except InvalidConfigurationError as e:
+        click.echo('Error: Cannot parse config file', err=True)
+        click.echo(f' - {e.message}', err=True)
+        sys.exit(1)
     except MissingConfigurationError as e:
         click.echo('Error: Missing configuration', err=True)
         for missing_item in e.missing: